# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pareeksha API",
    "description": "Student registration and auth API with role-based access (student, admin, public).",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development" }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "status": { "type": "string", "example": "ok" } } }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "description": "Returns JWT and user. Use the token in `Authorization: Bearer <token>` for protected routes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "admin@pareeksha.local" },
                  "password": { "type": "string", "example": "admin123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "token": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "email": { "type": "string" },
                        "role": { "type": "string", "enum": ["student", "admin", "public"] },
                        "studentId": { "type": "string", "nullable": true }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/students/register": {
      "post": {
        "summary": "Register student (public)",
        "description": "Creates a student record and a user with role `student`. No auth required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StudentRegistration" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "student": { "$ref": "#/components/schemas/StudentRecord" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "errors": {
                      "type": "array",
                      "items": { "type": "object", "properties": { "field": { "type": "string" }, "message": { "type": "string" } } }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/students/me": {
      "get": {
        "summary": "Get my profile (student)",
        "description": "Returns the authenticated student's own record. Requires role `student`.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "student": { "$ref": "#/components/schemas/StudentRecord" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/students": {
      "get": {
        "summary": "List all students (admin)",
        "description": "Returns all registered students. Requires role `admin`.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "students": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/StudentRecord" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/students/{id}": {
      "get": {
        "summary": "Get student by ID",
        "description": "Admin: any student. Student: own record only.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "student": { "$ref": "#/components/schemas/StudentRecord" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" }
        }
      },
      "StudentRegistration": {
        "type": "object",
        "required": [
          "preferredLanguage", "adhaarNumber", "firstName", "lastName", "password",
          "dateOfBirth", "gender", "schoolNameAndAddress", "schoolEnrollmentNumber",
          "class", "board", "addressLine1", "city", "state", "country", "pincode", "email", "mobile"
        ],
        "properties": {
          "preferredLanguage": { "type": "string", "example": "English" },
          "adhaarNumber": { "type": "string", "minLength": 12, "maxLength": 12, "example": "123456789012" },
          "firstName": { "type": "string" },
          "middleName": { "type": "string" },
          "lastName": { "type": "string" },
          "password": { "type": "string", "minLength": 6 },
          "dateOfBirth": { "type": "string", "format": "date" },
          "gender": { "type": "string", "enum": ["Male", "Female", "Other"] },
          "schoolNameAndAddress": { "type": "string" },
          "schoolEnrollmentNumber": { "type": "string" },
          "class": { "type": "string", "enum": ["1","2","3","4","5","6","7","8","9","10","11","12"] },
          "board": { "type": "string", "enum": ["CBSE", "ICSE", "State Board"] },
          "addressLine1": { "type": "string" },
          "addressLine2": { "type": "string" },
          "city": { "type": "string" },
          "state": { "type": "string" },
          "country": { "type": "string" },
          "pincode": { "type": "string", "minLength": 6, "maxLength": 6 },
          "email": { "type": "string", "format": "email" },
          "mobile": { "type": "string", "minLength": 10, "maxLength": 10 }
        }
      },
      "StudentRecord": {
        "allOf": [
          { "$ref": "#/components/schemas/StudentRegistration" },
          {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "createdAt": { "type": "string", "format": "date-time" }
            }
          }
        ]
      }
    }
  }
}